Add support for parsing symbolic L2 norm costs (#21394)
In a recent underactuated pset, we found that *lots* of people wanted
to pass e.g. np.linalg.norm(x1-x2) to the AddCost methods, and were
stymied by a lack of support there. Asking them to instead suddenly
understand how to create an L2NormCost, create the binding, and pass
_that_ to (GCS's version of) AddCost was too much.

This PR adds DecomposeL2NormExpression, and uses it in
prog.AddCost(expression). Consistent with the other Add*Cost methods,
it also adds prog.AddL2NormCost(expression), with all of the specific
tolerance arguments.