Removes DRAKE_DEMANDs temporily added during a PR train transition period. (#21469)
We added a number of DRAKE_DEMANDs in SapDriver::AddContactConstraints() during
a transition period in PR #20913. In that PR we explicitly loaded NANs to mark
not yet supported configurations (in this case deformables + HuntCrossley
contact constraints). Now that the feature is complete and those configurations
are supported, these DEMANDs can go away.