Parse glTF files into RenderEngineGl
The main work is done by the GltfMeshExtractor defined in
internal_render_engine_gl.cc. It extracts data from the glTF file and
directly inserts it into RenderEngineGl's various geometry and mesh caches
and informs the instantation of those canoncial meshes.

This also includes the ability to use textures that are defined in memory
instead of reading them from the disk (in support of embedded textures).

Further allowances are made for glTF peculiarities:

  - glTF defines image origin bottom left, but OpenGl defines them top
    left.
  - An embedded image is vertically flipped relative to the same image
    stored as an external file

Finally, update the supported file format documentation to provide
guidance on using glTF files with RenderEngineGl.