Reduce the noise in meshcat visualizer realtime rate reports
Adds a new API Meshcat realted to realtime rate. Rather than having clients
compute the realtime rate themselves (and immediately broadcasting it
to all clients), clients simply report to Meshcat that sim time has
advanced.

Meshcat uses the report of time advancement to throttle the report of real
time rate (and computes the real time rate itself based on the amount the
simulation has advanced since the last broadcast message.

Also changed the display of the realtime rate -- rather than streaming
values as fast as it can, it only changes when it receives a message doing
a better job of giving the user feedback on how frequently messages are
coming through.

Uses the new API to enable meldis RTR simply.