Move unit tests of boxes_overlap out of obb_test (#21546)
The code in boxes_overlap.cc was indirectly tested in obb_test.cc.
Now boxes_overlap_test.cc takes the responsibility for testing
the code directly (and obb_test.cc has become that much simpler).

This also includes a tweak to MaybePauseForUser() to include an
optional prompt message. The prompt message only gets printed if
the pause is enabled.