[multibody] Fix many dtors to not be inline
We audit and fix a bunch of destructors to be non-defaulted.

Fixing that also points out that some of our "nonsymbolic" docs were
lying -- we were compiling for symbolic even though we claimed it was
out of scope, so we also fix the docs / macros to add symbolic in a
few places.