[Mujoco parser] Don't call CalcSpatialInertia unless needed
Resolves #21648.

This defers the call to CalcSpatialInertia until/unless it is actually
needed. If inertia values are specified for the body in the xml file,
then we don't call CalcSpatialInertia. This allows us to successfully
parse models where CalcSpatialInertia fails on the obj, but the
calculation was not needed.

Adds the Rethink Robotics Sawyer (Apache 2.0) model to the
detail_mujoco_parser_test; parsing fails on this model without the fix
in this PR.