Fall back to Obb when CalcSpatialInertia fails.
This adds the realsense model (Apache 2.0) from the menagerie for the
test, and required additionally adding support for zero-sized
capsules. It was extremely hard to find an example in the menagerie
that failed for _only_ the CalcSpatialInertia
problem... hello_robot_stretch and others also failed for some
additional reason, so were not suitable for this test.