[Mujoco parser] Don't call CalcSpatialInertia unless needed (#21665)
Resolves #21648.

This defers the call to CalcSpatialInertia until/unless it is actually
needed. If inertia values are specified for the body in the xml file,
then we don't call CalcSpatialInertia. This allows us to successfully
parse models where CalcSpatialInertia fails on the obj, but the
calculation was not needed.

Adds the Rethink Robotics Sawyer (Apache 2.0) model and Hello Robot
Stretch (Clear BSD) to the detail_mujoco_parser_test. For Sawyer
parsing fails on this model without the fix in this PR. For Stretch,
parsing still fails, for a different reason -- it uses non-uniform
scaling on a mesh which we do not support yet -- but at least the
IsPhysicallyValid() errors have been resolved.