[tools] Switch to bazel's deps vs implementation_deps semantics (#21701)
At some point Bazel said that interface_deps vs deps was going to
be the way and the light for cc_library in the future, and then later
they changed their mind. As of today, deps are the interface deps and
implementation_deps are the implementation deps. Rewrite all of our
BUILD files and macros to match the new spelling.

Update buildifier hints for better precision.