Implements margin for convex hulls.
In addition to the new functionality, this PR also fixes a bug, a wrong index,
in OderPolyVertices() in make_convex_hull_impl.cc. The additional unit tests
introduced in the PR, sensitive to the behavior of OderPolyVertices(), provide
sufficient test coverage for the bug fix, i.e. the new tests do not succeed
without it.