Implement SamplePaths for GraphOfConvexSets (#21402)
Implements a function `SamplePaths` that uses the specified flow
variables to sample a number of random paths from a provided source to
a target vertex through the `GraphOfConvexSets` instance according to
the provided arguments. Returns a vector of potential paths from the
source to target (where each path is a vectors of edges). This
function refactors some of the current code in `SolveShortestPath`,
as well as addressing the bug described in 21662.