Introduce MemoryFile
Drake can load a file's contents into memory (where we keep its contents,
its extension, it's content's hash value, and some textual hint to the
user). The MemoryFile stores all of this.

It is explicitly used in the Meshcat FileStorage structure (but will be
used more in the future).

Bindings are included.

This includes an incidental improvement to reset_after_move giving
direct access to the managed value.