Bury the whole "unit density" thing inside InertiaCalculator
The old version returned a SpatialInertia with a mass that didn't have
the *meaning* of mass. This led to increased complexity of documentation
and made things weird for the caller.

Instead, the unit density has become an internal implementation trick and
the interface legitimately returns the quantities it returns, labeled as
such.