OBJ files for proximity can have multiple objects (#21745)
Currently, Mesh objects are (generally) replaced with their convex hull
in most proximity roles (with notable exceptions related to hydroelastic
contact).

Historically, there has been a requirement that the .obj file contain only
a single object. For this particular domain, the requirement is no longer
relevant, so we've removed it. This entails the following:

- Rename an .obj file to more directly reflect its new role (also
  stripping out geometry data that was unhelpful).
- Update the read_obj_test.cc to make the test less redundant and to
  eliminate some extant TODOs.