Update working with free bodies to reflect current implementations
Much of the free body documentation still assumes that all free bodies
have the world as parent. That's not generally true. In fact, we advocate
simulating things like suction by putting a floating joint between end
effector and manipuland and simply locking the joint when the end
effector is "holding" the manipuland. The configuration of the joint is
always relative to the inboard frame which, in this case, is not the
world.

This amends the documentation to distinguish between "free" bodies and
"floating base" bodies.

Furthermore, the API Mbp::SetFreeBodyPoseInWorldFrame() is a lie. For the
new terminology, it doesn't actually operate on "free" bodies -- only
"floating base" bodies. Its documentation has been corrected to reflect
the *true* state. While the name is still misleading, it is left intact
in case we can correct the implementation in the future to match the name.