Update working with free bodies to reflect current implementations
Much of the free body documentation still assumes that all free bodies
have the world as parent. That's not generally true. In fact, we advocate
simulating things like suction by putting a floating joint between end
effector and manipuland and simply locking the joint when the end
effector is "holding" the manipuland. The configuration of the joint is
always releative to the inboard frame which, in this case, is not the world.

Furthermore, the API Mbp::SetFreeBodyPoseInWorldFrame() doesn't serve its
purpose anymore because it doesn't perform the necessary kinematics.

TODO:
 - Fix SetFreeBodyPoseInWorldFrame()
  - Possibly just deprecate it.
 - Continue deprecating python bindings with misleading parameter names.
   - parameter names X_WB are no longer valid.
   - one deprecation for each function with changed parameter. (First in
     place).