Making Meldis and DrakeVisualizer in-memory mesh compatible
DrakeVisualizer encodes in-memory meshes as a literal json representation
of the InMemoryMesh (encoding into base64 conservatively).

The solar system example uses DrakeVisualizer (and therefore relies on
meldis). It is the ideal platform for testing meldis and an in-memory mesh.

By adding MeshcatVisualizer to the solar system example, it also lets
us confirm meshcat visualizer's compatibility with in-memory meshes.