Introduce InMemoryMesh and MeshSource
These provide the foundation of in-memory mesh specification.
InMemoryMesh tracks a MemoryFile for a mesh file (and an optional set of
supporting files in the near future).

MeshSource acts as a sugared-up variant between a filesystem::path and
an InMemoryMesh. This will, ultimately, serve as the storage for Mesh
and Convex specifications.

While it is not integrated with Mesh or Convex yet, we exercise the new
APIs in vtk_to_volume_mesh -- which consumes the MeshSource and creates
the VolumeMesh from both in-memory as well as on-disk .vtk files.