[GCS] Resolve inconsistencies in results from shortest path vs convex restriction (#21833)
**This is a breaking change** (Note that GCS is still marked experimental)

During the review of #21530, we realized that cost slack variables
were not being written to the MathematicalProgramResult in a
consistent way between SolveShortestPath and
SolveConvexRestriction. This PR addresses this by:

1) The "ell" slack variables associated with the AddCost methods are
not returned to the user. Users should call vertex->GetSolutionCost()
or edge->GetSolutionCost() to retrieve them. SolveConvexRestriction
doesn't use ell, and it's an implementation detail that shouldn't have
ever been exposed to the user.

2) GetSolutionCost and GetSolution are now more robust... if the
placeholder variables are not in the MathematicalProgramResult, they
return the implied values.

3) SolveConvexRestriction no longer tries to write NaN values for the
unused vertices on every solve. The more robust GetSolution methods
make this reasonable, and now that we've been generating large graphs
this has become more important. Users may need to change calls from
e.g. `result->GetSolution(v->x())` to `v->GetSolution(result)`.

4) The artificial `graphviz_options.show_costs = False` that was
needed in `trajectory_optimization_test.py` to cover up this issue is
now safely removed.

Resolves #20443.