Add FileSource
FileSource is a bit like MeshSource. How they differ is what the in-memory
representation is.

For a single file, a FileSource is either a path or a MemoryFile.
For mesh, a MeshSource is either a path, or a big group of files (each of
which is a file source, more or less).