Consolidate .obj parsing (#21842)
We had two different sets of functions that would parse *just* the
geometry out of an .obj file. They also had slightly different semantics.
There was no value in the separation.

In this case, we maintain the two APIs, but we have one API simply call
into the other API, limiting the spread of tinyobjloader.

There is a slight implication to the mesh parser. It previously enforced
that parsed meshes *must* have a single object. This has now been relaxed.
As the parsed geometry was only used to compute mass properties, this
relaxation seems harmless, and makes it more compatible with a larger set
of valid .obj files.