Mesh and Convex now uses MeshSource as storage
MeshSource replaces the "filename" and "extension" members. For now,
the InMemoryMesh-valued constructors are hidden from documentation and
filename() semantics is changed to throw if someone attempts to call
filename on an in-memory mesh. This throwing behavior is equivalent to
placing a `DRAKE_DEMAND(mesh.source().IsPath())` *everywhere* a
mesh.filename() is currently used. As support for MeshSource spreads,
errors will evaporate.

We exercise the new storage in a limited case: creating a rigid hydro mesh
from a MeshSource.

Furthermore, we update MakeVolumeMeshFromVtk:
 - That function should not be used on Convex -- we should be computing a
   VolumeMesh from its convex hull.
 - Rather than taking a path, it takes a Mesh and passes the mesh's source
   to ReadVtkToVolumeMesh().