Implement obj->mesh in terms of MeshSource
The obj-to-mesh utilities in read_obj.* and obj_to_surface_mesh.* have
been updated to take a MeshSource as an argument (relying on its implicit
conversion ability to maintain compatibility with call sites).

This had the following downstream effects:

 - detail_mesh_parser.cc was calling directly into DoReadObjToSurfaceMesh().
   Its invocation has changed to follow the new API.
 - Various error messages have evolved; tests attempting to capture the
   error messsages have followed suit.

To facilitate this effort, we patch tinyobjloader to facilitate parsing
from an in-memory file more efficiently.
  - The current API builds a stream from the data without copying.
  - A new API allows specifying a material reader instead of the mtl text.