RenderMesh and RenderMaterial parse .obj files from MeshSource
This builds on the addition of MeshSource having supporting files (as
material definitions require .mtl and possibly texture images).

The RenderMesh infrastructure includes a custom tinyobj::MaterialReader in
order to host .mtl file data from disk or memory.

This introduces a new internal class: TextureSource. This is akin to a
FileSource, but where FileSource has two modes, TextureSource has four
to span a greater number of definitions for glTFs embedded textures.

TODO:
 - Based on how conversation on `FileSource` goes, change IsKey to is_key,
   etc.