Extend convex hull computation to include in-memory glTF files
We use VTK to parse the glTF files. This moves from vtkGLTFImporter to
vtkGLTFDocumentLoader. This latter class supports parsing from a stream.
However, we do patch it so we can silence it regarding missing supporting
files that we simply don't care about (e.g., texture images).

To be able to parse from memory, we need a custom vtkURILoader to resolve
the file URI for a glTF's .bin file. We create one that knows about
MeshSource so that it can serve files to VTK appropriately.

To properly test a glTF file, we clean up rainbox_box.gltf, so that its
transforms more fully reflect the intended value (they'd been written out
with imprecision which was irrelevant to its visual appearance, but messed
up the convex hull).

In support of all of the above, we also do the following:

- Introduce vtk-based helper provides utilities to work with glTF files in
  VTK.
- Patching VTK
  - Add VTK::IOCore to our visible VTK packages.
  - The aforementioned patch.
- Make the VtkDiagnosticEventObserver visible to the convex hull
  computation.