Planar scene graph visualizer updated
All "mesh"-type objects (Mesh and Convex) ultimately are represented
by their convex hulls. This clarifies this fact. Furthermore, the
logic for searching for an alternative mesh file has been updated to
respect the MeshSource for the Mesh/Convex (we only search if the source
was a path and not in-memory).