[meshcat] when .mtl files are missing, render colors instead
Towards #20444. Like many mujoco assets, the
`mujoco_menagerie/boston_dynamics_spot/spot.xml` has obj files that
reference missing `.mtl` files. For `spot.xml`, the mujoco parser
correctly reads the rgba texture, but the meshcat parser did not
publish it (because it went down the branch expecting the mtl).

This change makes meshcat slightly more robust... if the mtl file is
missing, it publishes the file in the format that will accept the
color property instead.