openusd_internal: patch reduction progress (#21886)
Enable upgrade.py to work vs. the OpenUSD dev branch. This is useful for
testing upstreamable patches, since upstream's dev branch is where they
must be merged.

Since dev is quite a bit ahead of releases, we try to craft a build
system that can work equally well with both.