Add supporting files to InMemoryMesh (#21923)
This revisits the InMemoryMesh making it a simple struct, simplifying
bindings and call sites (and updating previously merged code).

Supporting files are implemented via a "FileSource" - an alias for a variant
between a filesystem::path and a drake::MemoryFile. This allows an
in-memory mesh's supporting files to reference resources on disk as well as
other in-memory files.