Avoid duplicate warnings in multibody XML parsers
The multibody parsers, in particular, had a habit of being very noisy,
since if an unsupported element or attribute is used once it is likely
used many times. This actually made it hard to visually inspect the
output to see the unique set of unsupported tags. This change
significantly improves the output for examples like in #20444.