[visualization] Fix meshcat control deletion paths
Fix a footgun that has been hiding in plain sight for some time: the
mismatch among the meshcat controls API, its (mis-)uses, and the fact
that Python bindings make destructor ordering unpredictable.

Why hasn't it fired? Largely because pybind memory management was
configured to leak diagrams and leaf systems forever, so destructors
would never run. In order to open the way to a fix for memory leaks, the
destructor hazard(s) must be addressed first.

* Introduce a non-strict deletion mode, that warns instead of throws.
* Use non-strict deletion when called from destructors.
* In addition, catch all exceptions in destructors, and abort if any are
  raised.
* Update bindings, doc, and tests.

Note that this patch does *not* solve any of the inherent hazards of the
meshcat controls API: the delusion of control ownership, problems with
overlapping client lifetimes, etc. It merely stops the destructor-operated
deletion paths from throwing.