Restore RenderEngineVtk MultiMaterialObject test for EGL
The crux of this was that rainbow_stripes.png has an alpha channel. When
we enabled shadows in VTK, we swapped out rendering pipelines which treated
meshes with those images differently. The test never really noticed that
these changed. Changing the backend merely caused the change to manifest
in a way that the test noticed (merely by happenstance).

The solution:
 - Remove the alpha channel.
 - Modify the test so it will be more sensitive to those kinds of changes.
   - Refactor the CompareImage code to use more widely (removing its
     behavior of saving the image to compare).
   - Make use of VTK's automatically save rendered image functionality.
   - Update callsites to CompareImage accordingly.

This does mean we don't have a test of transparent things. This is not a
problem:
  1. We only accidentally had one before.
  2. We don't support transparency with any kind of uniformity or
     reliability.
  3. We need a test that tracks what we *do* support and document
     supported transparency appropriately (a follow up PR).