[solvers] Port solver back-ends to use SpecificOptions (#21345)
In the near future, we anticipate changing the SolverOptions API in
support of loading and saving (i.e., serialization). That's especially
troublesome for our solver back-ends that consume its information,
given the low-level and hodge-podge ways in which they hunt for and
apply their specific options. In anticipation of that, here we port
all solvers to the new SolverBase::DoSolve2 virtual function that uses
SpecificOptions instead of SolverOptions. The DoSolve API remains
unchanged, for any out-of-tree solvers.