Correct VTK header generation from binary files (#22092)
The technique we were using to emulate vtkEncodeString.cmake was
insufficient for binary files. Up til now, we were only dependent on ascii
files and it was fine.

However, when using SSAO, we'll be dependent on a texture. That texture
doesn't get converted properly. The solution is to initialize the char
array as unsigned char with an array of one-byte hex values.