Add Expression Costs and Formula Constraints to Subgraphs in GcsTrajectoryOptimization (#22091)
Adds placeholder variables to subgraphs, that can be used to construct costs and constraints to each vertex or edge.

Note that this includes the fix  Fix a Bug in GCS When Passing in a Trivially-Infeasible Upper Bound #22090, since the conjunctive formula used in a test case creates an instance of this bug.

Co-Authored-By: Russ Tedrake <russt@mit.edu>