RenderEngineVtk can be forced into PBR mode
Adds a new parameter to RenderEngineVtkParams. Rather than waiting for the
engine's lighting model to get promoted based on its content, it can be
pushed into that mode upon construction.

This also fixes a bug in which cloned instances forgot whether or not
phong materials should be automatically promoted to PBR.