[framework] Update scalar conversion test utility
As of Xcode 16 or so, it is no longer safe to blithely dynamic_cast a
system to a type that it has statically forbidden conversion support
for. Instead, check (statically) for conversion support, and only then
step through the dynamic conversion test steps. Give a nice error
message if the conversion is forbidden statically.

While here, revert an obsolete compiler workaround, and fold the
implementation code down from two copies to one.