Fix a bug in MosekSolver on parsing PSD matrix with repeated variable. (#22173)
When one positive-semidefinite (PSD) matrix of symbolic variables contain a
repeated variable, with one of the entry being off-diagonal, the parsing was
 incorrect. For example, if we have a PSD matrix of
[x(0), x(1), x(2)]
[x(1), x(1), x(3)]
[x(2), x(3), x(4)]
where both the (0, 1) entry and (1, 1) entry contain the same variable,
and the (0, 1) entry is an off-diagonal entry, then we had a bug to
parse this PSD constraint. This PR fixes the bug.