Updates and documents MultibodyPlant's default proximity values. (#21463)
This PR updates the default value of proximity properties used by MbP.
The idea is that a user can create a model without specifying any contact parameters and run a first simulation with a reasonable approximation of rigid contact. For tighter "rigid contact" or furhter compliance, users of course must provide their contact parameters estimations. This workflow and guidelines for estimation are documented in a new module.

Label "release:fix" is related to the fact that these new contact parameters defaults affect the behavior of Drake's simulations.