[build] Use bzlmod for development and testing
The main change for bzlmod that we need to accomodate is that our
module name in runfiles is "_main" now instead of "drake", which
mostly affects some linters and tests that were hard-coded. More
generally, there is now a concept of "repo mapping" where module
names known to Drake's MODULE.bazel are projected to unique names
in runfiles, to allow for multiple copies of modules to co-exist.
This means that a runfiles lookup must not only specify the path
we're looking for, but from _whose point of view_ that path is
coming from.

* common: Adjust FindRunfile to know about source_repository. When
  none is given and we're seeking a drake runfile, supply the right
  "point of view" string automatically.

* resource_tool_test: Loosen test condition to be only the file
  basename.

* unittest_main: Adjust our source file scan to allow for either
  bzlmod or workspace spelling of the test program.

* tools: Adjust find_all_sources scan for .bazelproject to check
  against the realpath instead of the runfiles path.

As of this commit, all of our skylark dependencies have been ported
from WORKSPACE to bzlmod.

Note that even though Drake now uses bzlmod by default, it is not
yet possible for downstream projects to consume Drake as a bzlmod
module; that remains future work.