[build] Disable autoloaded (built-in) bazel rules
The release-notable changes here are the addition of rules_java and
rules_shell as new dependencies and the bump of the minimum supported
bazel version up to 7.4.

As part of uniform loading of rules we also refactor to provide a single
point of control for shell and java, like we had previously done for cc
and python and tidy up some missing pieces.