Adds a scene_graph() accessor to multibody::Parser.
Currently the Parser uses GetMutableSceneGraphPreFinalize() to access
the scene_graph, but per slack, this is not the long-term solution.

Nevertheless, conceptually the Parser _does_ have access to the
scene_graph (unless no scene_graph is registered) as well as the
plant, so adding this API should be durable even the
GetMutableSceneGraphPreFinalize() workflow goes away.