Fixes up InverseDynamicsController to use the actuation matrix (#22315)
Clarifies that the control output of the inverse dynamics controller
is actuation, not generalized forces, and uses the actuation matrix to
achieve this. Previously, by connecting the output of the
InverseDynamicsController system to an actuation input port of the
MultibodyPlant, we implicitly assumed that the actuator matrix was the
identity matrix. If the actuators were in a different order than the
joints (generalized forces), then our method produced the wrong
actuator inputs.

The generalized_force output port is still available, partly for
backwards compatibility, and because it's possible that there are
applications that prefer this.

This is a rewrite of PR #20971.