Yaml serialization includes std::filesystem::path and FileSource (#22288)
Serializable structs can now have members that are either
std::filesystem::path (or pathlib.Path) or drake::FileSource.

The FileSource support is limited. FileSource is a variant containing
either a std::filesystem::path or MemoryFile. If it contains a path
(its default configuration), serialization is fine. If it contains a
MemoryFile it throws. The throwing behavior will be updated later when
we've decided what we're going to do about bytestrings and the !!binary
yaml tag.