[build] Use bzlmod for development and testing (#22269)
In other words, Drake Developers will now be using bzlmod by default.
(CMake installs were already using it.)

Users who consume Drake as a bazel external must still use workspace
mode (no bzlmod). Even though Drake now uses bzlmod for developers,
it is not yet possible for downstream projects to consume Drake as a
bzlmod module; that remains future work.

The main change required for bzlmod here is that our module name in
runfiles is "_main" now instead of "drake", which mostly affects some
linters and tests that were hard-coded. More generally, there is now a
concept of "repo mapping" where module names in a MODULE.bazel file are
projected to unique names in runfiles, to allow for multiple copies of
modules to co-exist. This means that a runfiles lookup must not only
specify the path it's looking for, but from _whose point of view_ that
path is coming from.

* common: Adjust FindRunfile to know about source_repository. When
  none is given and we're seeking a drake runfile, supply the right
  "point of view" string automatically.

* resource_tool_test: Loosen test condition to be only the file
  basename.

* unittest_main: Adjust our source file scan to allow for either
  the bzlmod or workspace spelling of the test program. (The old
  workspace spelling would only be used by downstream projects.)

* tools: Adjust find_all_sources scan for .bazelproject to check
  against the realpath instead of the runfiles path.