[systems] Add DiagramBuilder overloads for shared_ptr
AddSystem and AddNamedSystem functions are overloaded to allow passing
the system via mutable shared_ptr (in addition to mutable unique_ptr).

Architectually, the diagram still maintains exclusive ownership of its
children: each child still has a back-pointer to the Diagram (via the
SystemParentServiceInterface) so it's not possible to actually "share"
a subsystem across multiple diagrams. However, unique_ptr is overly
restrictive: it not only denotes exclusive ownership (which is fine),
but also demands that resource cleanup is `delete MyClass` (which is
insufficient for our Python bindings). We must use a smart pointer
type that allows customizing cleanup with runtime data (i.e., uses
type erasure). We could imagine using unqiue_ptr whose Deleter was a
std::function, but it's simpler to just use shared_ptr.

Relatedly, we also adjust the Diagram subclasses PidControlledSystem
and RgbdSensorDiscrete to accept the more general shared_ptr (instead
of unique_ptr) for their child systems and to use ref_cycle for their
pydrake bindings.