PR [build] Remove two constants from python's version.bzl
These constants unnecessarily couple our BUILD files to our repository
rule, without the py_toolchain as an intermediary. Our goal is to use
the toolchain to fully encapsulate the selected python interpreter.

(1) PYTHON_EXTENSION_SUFFIX

Simplify how pydrake module shared libraries are named in our binary
packages. The importlib.machinery package defines many possible values
for valid EXTENSION_SUFFIXES (i.e., the filename extension used for
native code), e.g., on Ubuntu 22.04 the possible values are
['.cpython-310-x86_64-linux-gnu.so', '.abi3.so', '.so']. Previously we
used the first choice (with the python version baked in) but this is
awkward, so now we use the last option (plain '.so'). The distinction
would only matter if we packaged builds for multiple python versions
into the same release artifact at once, which we never do. The first
choice would also have become moot after we switch to using ABI 3 in
the future (in which case we'll use the middle choice).

Ditto for lcm-python's shared library.

(2) PYTHON_SITE_PACKAGES

This is no longer provided as an opaque constant; its correct value is
trivially derivable from the Python version number so we'll just use
the longhand anywhere we need such a path.

While we're anyway cleaning up the repository.bzl, we also inline the
"interpreter_attrs" dictionary because its name was a lie and keeping
it out-of-line doesn't gain us anything anyway.