[workspace] Remove unnecessary nlopt_internal patch
Upstream uses target_compile_definitions(PRIVATE) to propagate this
setting. When not enabled, the pre-processor definition is simply
absent and that's entirely consistent with its use of '#ifdef' in
optimize.c (vs '#if' if it were a header file setting that was
expected to always either be 0 or 1). There's no need to feed the
setting into the header file.