Change gibberish string to recongizable string
The bytes (10, 20, 30) get severely mangled when printed as a hex string of
escaped characters (\n\x14\x1e). Given the actual byte values never appear
in tests and only get seen as the gibberish, it makes sense to swap one
set of random-seeming gibberish with an alternative, human-recognizable
pattern of gibberish.