[systems] LuenbergerObserver and KalmanFilter use shared_ptr (#22351)
The LuenbergerObserver and KalmanFilter now take the observed system
as a shared_ptr-to-const, not unique_ptr-to-mutable. The functions
don't need to mutate anything and don't need exclusive ownership, and
passing by unique_ptr causes trouble in our pydrake bindings.

Overload SteadyStateKalmanFilter to accept the context by-const-ref
instead of the unnecessary by-mutable-unique_ptr.