Improve how we manage which PDM to use
Move our PDM version pin to a separate file. This both makes it easier
to manage updates to the same and also allows the pinned version to be
usable outside of the venv_sync script. Add a simple script to update
version pins in this file. Teach venv_upgrade to use this to update the
PDM version pin in addition to updating the PDM lock file. Also, modify
venv_sync to tell PDM to not complain about newer versions.