[trajectories] Deprecate overriding public virtual functions (#22395)
Subclasses must switch to using NVI overrides, instead.

In most cases where subclass functions moved from public to private we
also relocate the code in the cc file to stay in sync with header file
order. However, the piecewise_polynomial.cc file is so disorganized
that we declare bankruptcy and don't move anything around in there.