Use PDM to manage macOS wheel dependencies
Use Drake's PDM-managed virtual environment (@python), rather than a
separate virtual environment, to manage the additional Python bits that
are needed to build a wheel. This allows us to pin the versions of these
dependencies, for better reproducibility and is one less place where
such things are being managed. Having the wheel dependencies installed
"up front" is also in line with how Linux wheels are built.