Modify install_prereqs to create a venv
Modify the install_prereqs script used for Drake tarball installations
to create a Python virtual environment in the location where Drake has
been extracted. On macOS, also use this to install Drake's Python
dependencies using PDM. Update user documentation accordingly.

The changes for Ubuntu are mainly to preserve consistency. For macOS,
this change means we are now using the pinned versions of our
dependencies and are no longer keeping a duplicate copy of this list
solely for installations.