[mujoco parser] upgrade unsupported mesh to error
Previously if a mesh was unsupported, then we only produceed a warning
and ignored the associated geometry. Now it is an error. (We can use
make_drake_compatible_mesh in the manipulation repo to batch convert
stl to obj, etc).

Also removes the "if you said stl but I can find a juxtaposed obj,
I'll use it" logic.

The non-uniform mesh warning was turned to an error (non-uniform
meshes already threw an error, but later in the pipeline).

All of the errors now reference
http://drake.mit.edu/troubleshooting.html which points to
make_drake_compatible_model.

Resolves #20444.  Hoorah!